made changes
made changes To BlindClass and Blind Adapter to get data to display
correctly,  fixed the Timer next and previous methods.  Delete methods
on both Blinds and Tournaments work within the application but the
changes from the delete does not save to disk, so the data loaded is
currently being added from the save.